pyproject.toml
[tool.poetry]
name = "fastapi-todo"
version = "0.1.0"
description = "FastAPI example using DI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.1"
uvicorn = { extras = ["standard"], version = "^0.16.0" }
dependency-injector = "^4.37.0"
redis = "^4.1.1"
"dogpile.cache" = "^1.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
